[ENH] Bundle CLI in python package
