+++ rust/cli/src/commands/run.rs
@@ -36,6 +36,9 @@ pub struct RunArgs {
         help = "The port to run the server on"
     )]
     port: Option<u16>,
+
+    #[clap(long, hide = true)]
+    test: Option<bool>,
 }
 
 fn validate_host(address: &String, port: u16) -> bool {
@@ -86,6 +89,8 @@ fn display_run_message(config: &FrontendServerConfig) {
 }
 
 pub fn run(args: RunArgs) {
+    let test = args.test.unwrap_or(false);
+    
     let config = match &args.config_path {
         Some(config_path) => {
             if !std::path::Path::new(config_path).exists() {
@@ -104,6 +109,10 @@ pub fn run(args: RunArgs) {
     };
 
     display_run_message(&config);
+    
+    if test {
+        return;
+    }
 
     let runtime = tokio::runtime::Runtime::new().expect("Failed to start Chroma");
     runtime.block_on(async {
