+++ rust/cli/src/lib.rs
@@ -1,0 +1,43 @@
+mod commands;
+mod utils;
+
+use crate::commands::run::{run, RunArgs};
+use clap::{Parser, Subcommand};
+
+#[derive(Subcommand, Debug)]
+enum Command {
+    Docs,
+    Run(RunArgs),
+    Support,
+}
+
+#[derive(Parser, Debug)]
+#[command(name = "chroma")]
+#[command(version = "1.0.0")]
+#[command(about = "A CLI for Chroma", long_about = None)]
+struct Cli {
+    #[command(subcommand)]
+    command: Command,
+}
+
+pub fn chroma_cli(args: Vec<String>) {
+    let cli = Cli::parse_from(args);
+
+    match cli.command {
+        Command::Docs => {
+            let url = "https://docs.trychroma.com";
+            if webbrowser::open(url).is_err() {
+                eprintln!("Error: Failed to open the browser. Visit {}.", url);
+            }
+        }
+        Command::Run(args) => {
+            run(args);
+        }
+        Command::Support => {
+            let url = "https://discord.gg/MMeYNTmh3x";
+            if webbrowser::open(url).is_err() {
+                eprintln!("Error: Failed to open the browser. Visit {}.", url);
+            }
+        }
+    }
+}
