+++ rust/cli/src/main.rs
@@ -1,42 +1,9 @@
-mod commands;
-mod utils;
-use crate::commands::run::{run, RunArgs};
-use clap::{Parser, Subcommand};
+#![windows_subsystem = "console"]
 
-#[derive(Subcommand, Debug)]
-enum Command {
-    Docs,
-    Run(RunArgs),
-    Support,
-}
-
-#[derive(Parser, Debug)]
-#[command(name = "chroma")]
-#[command(version = "0.0.1")]
-#[command(about = "A CLI for Chroma", long_about = None)]
-struct Cli {
-    #[command(subcommand)]
-    command: Command,
-}
+use std::env;
+use chroma_cli::chroma_cli;
 
 fn main() {
-    let cli = Cli::parse();
-
-    match cli.command {
-        Command::Docs => {
-            let url = "https://docs.trychroma.com";
-            if webbrowser::open(url).is_err() {
-                eprintln!("Error: Failed to open the browser. Visit {}.", url);
-            }
-        }
-        Command::Run(args) => {
-            run(args);
-        }
-        Command::Support => {
-            let url = "https://discord.gg/MMeYNTmh3x";
-            if webbrowser::open(url).is_err() {
-                eprintln!("Error: Failed to open the browser. Visit {}.", url);
-            }
-        }
-    }
+    let args: Vec<String> = env::args().collect();
+    chroma_cli(args)
 }
