+++ rust/python_bindings/Cargo.toml
@@ -28,5 +28,6 @@ chroma-cache = { workspace = true }
 chroma-system = { workspace = true }
 chroma-types = { workspace = true, features = ["pyo3"] }
 chroma-error = { workspace = true, features = ["validator"] }
+chroma-cli = { workspace = true }
 mdac = { workspace = true }
