+++ rust/python_bindings/src/bindings.rs
@@ -22,8 +22,9 @@ use chroma_types::{
     ListCollectionsRequest, ListDatabasesRequest, Metadata, QueryResponse, UpdateCollectionRequest,
     UpdateMetadata,
 };
-use pyo3::{exceptions::PyValueError, pyclass, pymethods, types::PyAnyMethods, PyObject, Python};
+use pyo3::{exceptions::PyValueError, pyclass, pyfunction, pymethods, types::PyAnyMethods, PyObject, Python};
 use std::time::SystemTime;
+use chroma_cli::chroma_cli;
 
 const DEFAULT_DATABASE: &str = "default_database";
 const DEFAULT_TENANT: &str = "default_tenant";
@@ -46,6 +47,20 @@ impl PythonBindingsConfig {
     pub fn py_new(sqlite_db_config: SqliteDBConfig) -> Self {
         PythonBindingsConfig { sqlite_db_config }
     }
+}
+
+#[pyfunction]
+#[pyo3(signature = (py_args=None))]
+#[allow(dead_code)]
+pub fn cli(py_args: Option<Vec<String>>) -> ChromaPyResult<()> {
+    let args = py_args.unwrap_or_else(|| std::env::args().collect());
+    let args = if args.is_empty() {
+        vec!["chroma".to_string()]
+    } else {
+        args
+    };
+    chroma_cli(args);
+    Ok(())
 }
 
 //////////////////////// PyMethods Implementation ////////////////////////
